+++ ios/secure_app_switcher/Package.swift
@@ -5,7 +5,7 @@ import PackageDescription
 let package = Package(
     name: "secure_app_switcher",
     platforms: [
-        .iOS(.v9)
+        .iOS(.v13)
     ],
     products: [
         .library(