+++ lib/secure_app_switcher.dart
@@ -1,4 +1,2 @@
-library secure_app_switcher;
-
 export 'src/secure_app_switcher.dart';
 export 'src/secure_app_switcher_page.dart';